+++ aea/skills/base.py
@@ -25,11 +25,11 @@ import logging
 import queue
 import re
 from abc import ABC, abstractmethod
-from logging import Logger
+from logging import Logger, LoggerAdapter
 from pathlib import Path
 from queue import Queue
 from types import SimpleNamespace
-from typing import Any, Dict, Optional, Sequence, Set, Tuple, Type, cast
+from typing import Any, Dict, Optional, Sequence, Set, Tuple, Type, Union, cast
 
 from aea.components.base import Component
 from aea.configurations.base import (
@@ -45,6 +45,7 @@ from aea.context.base import AgentContext
 from aea.contracts.base import Contract
 from aea.exceptions import AEAException
 from aea.helpers.base import load_aea_package, load_module
+from aea.helpers.logging import AgentLoggerAdapter
 from aea.mail.base import Address
 from aea.multiplexer import OutBox
 from aea.protocols.base import Message
@@ -73,16 +74,16 @@ class SkillContext:
 
         self._is_active = True  # type: bool
         self._new_behaviours_queue = queue.Queue()  # type: Queue
-        self._logger = None  # type: Optional[Logger]
+        self._logger: Optional[Union[Logger, LoggerAdapter]] = None
 
     @property
-    def logger(self) -> Logger:
+    def logger(self) -> Union[Logger, LoggerAdapter]:
         """Get the logger."""
         assert self._logger is not None, "Logger not set."
         return self._logger
 
     @logger.setter
-    def logger(self, logger_: Logger) -> None:
+    def logger(self, logger_: Union[Logger, AgentLoggerAdapter]) -> None:
         assert self._logger is None, "Logger already set."
         self._logger = logger_
 
@@ -683,6 +684,21 @@ class Skill(Component):
         configuration.directory = Path(directory)
         return Skill.from_config(configuration, agent_context)
 
+    @property
+    def logger(self) -> Union[Logger, LoggerAdapter]:
+        """
+        Get the logger.
+
+        In the case of a skill, return the
+        logger provided by the skill context.
+        """
+        return self.skill_context.logger
+
+    @logger.setter
+    def logger(self, *args) -> None:
+        """Set the logger."""
+        raise ValueError("Cannot set logger to a skill component..")
+
     @classmethod
     def from_config(
         cls, configuration: SkillConfig, agent_context: AgentContext
@@ -704,7 +720,10 @@ class Skill(Component):
         skill_context = SkillContext()
         skill_context.set_agent_context(agent_context)
         logger_name = f"aea.packages.{configuration.author}.skills.{configuration.name}"
-        skill_context.logger = logging.getLogger(logger_name)
+        logger = AgentLoggerAdapter(
+            logging.getLogger(logger_name), agent_context.agent_name
+        )
+        skill_context.logger = logger
 
         skill = Skill(configuration, skill_context)