+++ packages/fetchai/skills/erc1155_client/handlers.py
@@ -47,7 +47,7 @@ from packages.fetchai.skills.erc1155_client.dialogues import (
 )
 from packages.fetchai.skills.erc1155_client.strategy import Strategy
 
-LEDGER_API_ADDRESS = "fetchai/ledger:0.1.0"
+LEDGER_API_ADDRESS = "fetchai/ledger:0.2.0"
 
 
 class FipaHandler(Handler):