+++ tests/test_cli/test_run.py
@@ -75,7 +75,7 @@ def test_run():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
 
@@ -86,7 +86,7 @@ def test_run():
             "config",
             "set",
             "agent.default_connection",
-            "fetchai/http_client:0.4.0",
+            "fetchai/http_client:0.5.0",
         ],
     )
     assert result.exit_code == 0
@@ -167,9 +167,9 @@ def test_run_with_default_connection():
 @pytest.mark.parametrize(
     argnames=["connection_ids"],
     argvalues=[
-        ["fetchai/http_client:0.4.0,{}".format(str(DEFAULT_CONNECTION))],
-        ["'fetchai/http_client:0.4.0, {}'".format(str(DEFAULT_CONNECTION))],
-        ["fetchai/http_client:0.4.0,,{},".format(str(DEFAULT_CONNECTION))],
+        ["fetchai/http_client:0.5.0,{}".format(str(DEFAULT_CONNECTION))],
+        ["'fetchai/http_client:0.5.0, {}'".format(str(DEFAULT_CONNECTION))],
+        ["fetchai/http_client:0.5.0,,{},".format(str(DEFAULT_CONNECTION))],
     ],
 )
 def test_run_multiple_connections(connection_ids):
@@ -194,7 +194,7 @@ def test_run_multiple_connections(connection_ids):
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
 
@@ -252,7 +252,7 @@ def test_run_unknown_private_key():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
     result = runner.invoke(
@@ -262,7 +262,7 @@ def test_run_unknown_private_key():
             "config",
             "set",
             "agent.default_connection",
-            "fetchai/http_client:0.4.0",
+            "fetchai/http_client:0.5.0",
         ],
     )
     assert result.exit_code == 0
@@ -291,7 +291,7 @@ def test_run_unknown_private_key():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.5.0"],
         standalone_mode=False,
     )
 
@@ -328,7 +328,7 @@ def test_run_unknown_private_key():
 
 #     result = runner.invoke(
 #         cli,
-#         [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+#         [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
 #     )
 #     assert result.exit_code == 0
 #     result = runner.invoke(
@@ -338,7 +338,7 @@ def test_run_unknown_private_key():
 #             "config",
 #             "set",
 #             "agent.default_connection",
-#             "fetchai/http_client:0.4.0",
+#             "fetchai/http_client:0.5.0",
 #         ],
 #     )
 #     assert result.exit_code == 0
@@ -367,7 +367,7 @@ def test_run_unknown_private_key():
 
 #     result = runner.invoke(
 #         cli,
-#         [*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.4.0"],
+#         [*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.5.0"],
 #         standalone_mode=False,
 #     )
 
@@ -403,7 +403,7 @@ def test_run_fet_private_key_config():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
 
@@ -427,7 +427,7 @@ def test_run_fet_private_key_config():
 
     error_msg = ""
     try:
-        cli.main([*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.4.0"])
+        cli.main([*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.5.0"])
     except SystemExit as e:
         error_msg = str(e)
 
@@ -462,7 +462,7 @@ def test_run_ethereum_private_key_config():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
 
@@ -486,7 +486,7 @@ def test_run_ethereum_private_key_config():
 
     error_msg = ""
     try:
-        cli.main([*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.4.0"])
+        cli.main([*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.5.0"])
     except SystemExit as e:
         error_msg = str(e)
 
@@ -522,7 +522,7 @@ def test_run_ledger_apis():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
     result = runner.invoke(
@@ -532,7 +532,7 @@ def test_run_ledger_apis():
             "config",
             "set",
             "agent.default_connection",
-            "fetchai/http_client:0.4.0",
+            "fetchai/http_client:0.5.0",
         ],
     )
     assert result.exit_code == 0
@@ -569,7 +569,7 @@ def test_run_ledger_apis():
                 "aea.cli",
                 "run",
                 "--connections",
-                "fetchai/http_client:0.4.0",
+                "fetchai/http_client:0.5.0",
             ],
             stdout=subprocess.PIPE,
             env=os.environ.copy(),
@@ -618,7 +618,7 @@ def test_run_fet_ledger_apis():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
     result = runner.invoke(
@@ -628,7 +628,7 @@ def test_run_fet_ledger_apis():
             "config",
             "set",
             "agent.default_connection",
-            "fetchai/http_client:0.4.0",
+            "fetchai/http_client:0.5.0",
         ],
     )
     assert result.exit_code == 0
@@ -662,7 +662,7 @@ def test_run_fet_ledger_apis():
                 "aea.cli",
                 "run",
                 "--connections",
-                "fetchai/http_client:0.4.0",
+                "fetchai/http_client:0.5.0",
             ],
             stdout=subprocess.PIPE,
             env=os.environ.copy(),
@@ -712,7 +712,7 @@ def test_run_with_install_deps():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
     result = runner.invoke(
@@ -722,7 +722,7 @@ def test_run_with_install_deps():
             "config",
             "set",
             "agent.default_connection",
-            "fetchai/http_client:0.4.0",
+            "fetchai/http_client:0.5.0",
         ],
     )
     assert result.exit_code == 0
@@ -738,7 +738,7 @@ def test_run_with_install_deps():
                 "run",
                 "--install-deps",
                 "--connections",
-                "fetchai/http_client:0.4.0",
+                "fetchai/http_client:0.5.0",
             ],
             env=os.environ,
             maxread=10000,
@@ -784,7 +784,7 @@ def test_run_with_install_deps_and_requirement_file():
 
     result = runner.invoke(
         cli,
-        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.4.0"],
+        [*CLI_LOG_OPTION, "add", "--local", "connection", "fetchai/http_client:0.5.0"],
     )
     assert result.exit_code == 0
     result = runner.invoke(
@@ -794,7 +794,7 @@ def test_run_with_install_deps_and_requirement_file():
             "config",
             "set",
             "agent.default_connection",
-            "fetchai/http_client:0.4.0",
+            "fetchai/http_client:0.5.0",
         ],
     )
     assert result.exit_code == 0
@@ -814,7 +814,7 @@ def test_run_with_install_deps_and_requirement_file():
                 "run",
                 "--install-deps",
                 "--connections",
-                "fetchai/http_client:0.4.0",
+                "fetchai/http_client:0.5.0",
             ],
             env=os.environ,
             maxread=10000,
@@ -872,7 +872,7 @@ class TestRunFailsWhenExceptionOccursInSkill:
                 "add",
                 "--local",
                 "connection",
-                "fetchai/http_client:0.4.0",
+                "fetchai/http_client:0.5.0",
             ],
             standalone_mode=False,
         )
@@ -889,7 +889,7 @@ class TestRunFailsWhenExceptionOccursInSkill:
 
         try:
             cli.main(
-                [*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.4.0"]
+                [*CLI_LOG_OPTION, "run", "--connections", "fetchai/http_client:0.5.0"]
             )
         except SystemExit as e:
             cls.exit_code = e.code
@@ -1084,7 +1084,7 @@ class TestRunFailsWhenConnectionConfigFileNotFound:
         """Set the test up."""
         cls.runner = CliRunner()
         cls.agent_name = "myagent"
-        cls.connection_id = PublicId.from_str("fetchai/http_client:0.4.0")
+        cls.connection_id = PublicId.from_str("fetchai/http_client:0.5.0")
         cls.connection_name = cls.connection_id.name
         cls.connection_author = cls.connection_id.author
         cls.cwd = os.getcwd()
@@ -1118,7 +1118,7 @@ class TestRunFailsWhenConnectionConfigFileNotFound:
                 "config",
                 "set",
                 "agent.default_connection",
-                "fetchai/http_client:0.4.0",
+                "fetchai/http_client:0.5.0",
             ],
         )
         assert result.exit_code == 0
@@ -1177,7 +1177,7 @@ class TestRunFailsWhenConnectionNotComplete:
         """Set the test up."""
         cls.runner = CliRunner()
         cls.agent_name = "myagent"
-        cls.connection_id = PublicId.from_str("fetchai/http_client:0.4.0")
+        cls.connection_id = PublicId.from_str("fetchai/http_client:0.5.0")
         cls.connection_author = cls.connection_id.author
         cls.connection_name = cls.connection_id.name
         cls.cwd = os.getcwd()
@@ -1211,7 +1211,7 @@ class TestRunFailsWhenConnectionNotComplete:
                 "config",
                 "set",
                 "agent.default_connection",
-                "fetchai/http_client:0.4.0",
+                "fetchai/http_client:0.5.0",
             ],
         )
         assert result.exit_code == 0
@@ -1269,7 +1269,7 @@ class TestRunFailsWhenConnectionClassNotPresent:
         """Set the test up."""
         cls.runner = CliRunner()
         cls.agent_name = "myagent"
-        cls.connection_id = "fetchai/http_client:0.4.0"
+        cls.connection_id = "fetchai/http_client:0.5.0"
         cls.connection_name = "http_client"
         cls.cwd = os.getcwd()
         cls.t = tempfile.mkdtemp()
@@ -1302,7 +1302,7 @@ class TestRunFailsWhenConnectionClassNotPresent:
                 "config",
                 "set",
                 "agent.default_connection",
-                "fetchai/http_client:0.4.0",
+                "fetchai/http_client:0.5.0",
             ],
         )
         assert result.exit_code == 0