+++ tests/test_cli_gui/test_misc.py
@@ -39,7 +39,7 @@ def test_home_page_exits():
 
     # assert the status code of the response
     assert result.status_code == 200
-    assert "AEA CLI REST API" in str(result.data)
+    assert "AEA GUI" in str(result.data)
 
 
 def test_icon():