+++ tests/test_packages/test_skills/test_carpark.py
@@ -19,9 +19,11 @@
 
 """This test module contains the integration test for the weather skills."""
 
+import pytest
+
 from aea.test_tools.test_cases import AEATestCaseMany, UseOef
 
-from tests.conftest import FUNDED_FET_PRIVATE_KEY_1
+from tests.conftest import FUNDED_FET_PRIVATE_KEY_1, MAX_FLAKY_RERUNS
 
 
 class TestCarPark(AEATestCaseMany, UseOef):
@@ -33,14 +35,14 @@ class TestCarPark(AEATestCaseMany, UseOef):
         carpark_client_aea_name = "my_carpark_client_aea"
         self.create_agents(carpark_aea_name, carpark_client_aea_name)
 
-        default_routing = {"fetchai/ledger_api:0.1.0": "fetchai/ledger:0.1.0"}
+        default_routing = {"fetchai/ledger_api:0.1.0": "fetchai/ledger:0.2.0"}
 
         # Setup agent one
         self.set_agent_context(carpark_aea_name)
-        self.add_item("connection", "fetchai/oef:0.5.0")
-        self.add_item("connection", "fetchai/ledger:0.1.0")
-        self.add_item("skill", "fetchai/carpark_detection:0.5.0")
-        self.set_config("agent.default_connection", "fetchai/oef:0.5.0")
+        self.add_item("connection", "fetchai/oef:0.6.0")
+        self.add_item("connection", "fetchai/ledger:0.2.0")
+        self.add_item("skill", "fetchai/carpark_detection:0.6.0")
+        self.set_config("agent.default_connection", "fetchai/oef:0.6.0")
         setting_path = (
             "vendor.fetchai.skills.carpark_detection.models.strategy.args.is_ledger_tx"
         )
@@ -51,10 +53,10 @@ class TestCarPark(AEATestCaseMany, UseOef):
 
         # Setup agent two
         self.set_agent_context(carpark_client_aea_name)
-        self.add_item("connection", "fetchai/oef:0.5.0")
-        self.add_item("connection", "fetchai/ledger:0.1.0")
-        self.add_item("skill", "fetchai/carpark_client:0.5.0")
-        self.set_config("agent.default_connection", "fetchai/oef:0.5.0")
+        self.add_item("connection", "fetchai/oef:0.6.0")
+        self.add_item("connection", "fetchai/ledger:0.2.0")
+        self.add_item("skill", "fetchai/carpark_client:0.6.0")
+        self.set_config("agent.default_connection", "fetchai/oef:0.6.0")
         setting_path = (
             "vendor.fetchai.skills.carpark_client.models.strategy.args.is_ledger_tx"
         )
@@ -113,29 +115,30 @@ class TestCarPark(AEATestCaseMany, UseOef):
 class TestCarParkFetchaiLedger(AEATestCaseMany, UseOef):
     """Test that carpark skills work."""
 
+    @pytest.mark.flaky(reruns=MAX_FLAKY_RERUNS)  # cause possible network issues
     def test_carpark(self):
         """Run the weather skills sequence."""
         carpark_aea_name = "my_carpark_aea"
         carpark_client_aea_name = "my_carpark_client_aea"
         self.create_agents(carpark_aea_name, carpark_client_aea_name)
 
-        default_routing = {"fetchai/ledger_api:0.1.0": "fetchai/ledger:0.1.0"}
+        default_routing = {"fetchai/ledger_api:0.1.0": "fetchai/ledger:0.2.0"}
 
         ledger_apis = {"fetchai": {"network": "testnet"}}
 
         # Setup agent one
         self.set_agent_context(carpark_aea_name)
         self.force_set_config("agent.ledger_apis", ledger_apis)
-        self.add_item("connection", "fetchai/oef:0.5.0")
-        self.add_item("connection", "fetchai/ledger:0.1.0")
-        self.add_item("skill", "fetchai/carpark_detection:0.5.0")
-        self.set_config("agent.default_connection", "fetchai/oef:0.5.0")
+        self.add_item("connection", "fetchai/oef:0.6.0")
+        self.add_item("connection", "fetchai/ledger:0.2.0")
+        self.add_item("skill", "fetchai/carpark_detection:0.6.0")
+        self.set_config("agent.default_connection", "fetchai/oef:0.6.0")
         setting_path = "agent.default_routing"
         self.force_set_config(setting_path, default_routing)
         self.run_install()
 
         diff = self.difference_to_fetched_agent(
-            "fetchai/car_detector:0.6.0", carpark_aea_name
+            "fetchai/car_detector:0.7.0", carpark_aea_name
         )
         assert (
             diff == []
@@ -144,16 +147,16 @@ class TestCarParkFetchaiLedger(AEATestCaseMany, UseOef):
         # Setup agent two
         self.set_agent_context(carpark_client_aea_name)
         self.force_set_config("agent.ledger_apis", ledger_apis)
-        self.add_item("connection", "fetchai/oef:0.5.0")
-        self.add_item("connection", "fetchai/ledger:0.1.0")
-        self.add_item("skill", "fetchai/carpark_client:0.5.0")
-        self.set_config("agent.default_connection", "fetchai/oef:0.5.0")
+        self.add_item("connection", "fetchai/oef:0.6.0")
+        self.add_item("connection", "fetchai/ledger:0.2.0")
+        self.add_item("skill", "fetchai/carpark_client:0.6.0")
+        self.set_config("agent.default_connection", "fetchai/oef:0.6.0")
         setting_path = "agent.default_routing"
         self.force_set_config(setting_path, default_routing)
         self.run_install()
 
         diff = self.difference_to_fetched_agent(
-            "fetchai/car_data_buyer:0.6.0", carpark_client_aea_name
+            "fetchai/car_data_buyer:0.7.0", carpark_client_aea_name
         )
         assert (
             diff == []