+++ tests/test_packages/test_skills/test_erc1155.py
@@ -52,23 +52,23 @@ class TestERCSkillsEthereumLedger(AEATestCaseMany, UseOef):
         }
         setting_path = "agent.ledger_apis"
         default_routing = {
-            "fetchai/ledger_api:0.1.0": "fetchai/ledger:0.1.0",
-            "fetchai/contract_api:0.1.0": "fetchai/ledger:0.1.0",
+            "fetchai/ledger_api:0.1.0": "fetchai/ledger:0.2.0",
+            "fetchai/contract_api:0.1.0": "fetchai/ledger:0.2.0",
         }
 
         # add packages for agent one
         self.set_agent_context(deploy_aea_name)
         self.force_set_config(setting_path, ledger_apis)
-        self.add_item("connection", "fetchai/oef:0.5.0")
-        self.add_item("connection", "fetchai/ledger:0.1.0")
-        self.set_config("agent.default_connection", "fetchai/oef:0.5.0")
+        self.add_item("connection", "fetchai/oef:0.6.0")
+        self.add_item("connection", "fetchai/ledger:0.2.0")
+        self.set_config("agent.default_connection", "fetchai/oef:0.6.0")
         self.set_config("agent.default_ledger", "ethereum")
         setting_path = "agent.default_routing"
         self.force_set_config(setting_path, default_routing)
-        self.add_item("skill", "fetchai/erc1155_deploy:0.7.0")
+        self.add_item("skill", "fetchai/erc1155_deploy:0.8.0")
 
         diff = self.difference_to_fetched_agent(
-            "fetchai/erc1155_deployer:0.7.0", deploy_aea_name
+            "fetchai/erc1155_deployer:0.8.0", deploy_aea_name
         )
         assert (
             diff == []
@@ -88,9 +88,9 @@ class TestERCSkillsEthereumLedger(AEATestCaseMany, UseOef):
         setting_path = "agent.ledger_apis"
         self.set_agent_context(client_aea_name)
         self.force_set_config(setting_path, ledger_apis)
-        self.add_item("connection", "fetchai/oef:0.5.0")
-        self.add_item("connection", "fetchai/ledger:0.1.0")
-        self.set_config("agent.default_connection", "fetchai/oef:0.5.0")
+        self.add_item("connection", "fetchai/oef:0.6.0")
+        self.add_item("connection", "fetchai/ledger:0.2.0")
+        self.set_config("agent.default_connection", "fetchai/oef:0.6.0")
         self.set_config("agent.default_ledger", "ethereum")
         setting_path = "agent.default_routing"
         self.force_set_config(setting_path, default_routing)